requirements.txt
streamlit
tensorflow
requests
pillow
